+++ spark-bigquery-dsv2/spark-bigquery-dsv2-common/src/main/java/com/google/cloud/spark/bigquery/v2/context/BigQueryDataSourceReaderContext.java
@@ -427,7 +427,8 @@ public class BigQueryDataSourceReaderContext {
   }
 
   public StatisticsContext estimateStatistics() {
-    if (table.getDefinition().getType() == TableDefinition.Type.TABLE) {
+    boolean isBigLakeManagedTable = BigQueryUtil.isBigLakeManagedTable(table);
+    if (BigQueryUtil.isBigQueryNativeTable(table)) {
       // Create StatisticsContext with information from read session response.
       final long tableSizeInBytes;
       final long numRowsInTable;
@@ -455,11 +456,13 @@ public class BigQueryDataSourceReaderContext {
           };
 
       return tableStatisticsContext;
-    } else if (table.getDefinition().getType() == TableDefinition.Type.EXTERNAL) {
+    } else if (table.getDefinition().getType() == TableDefinition.Type.EXTERNAL
+        || isBigLakeManagedTable) {
       ReadSession readSession = readSessionResponse.get().getReadSession();
       // Physical file size for BigLake tables is the size of the files post file pruning and
       // includes all fields.
       final long tablePhysicalSizeInBytes = readSession.getEstimatedTotalPhysicalFileSize();
+      final long tableRowCount = readSession.getEstimatedRowCount();
       final long originalRowSize = getRowSize(fields.values());
       final long projectedRowSize =
           schema.map(schema -> getRowSize(Arrays.asList(schema.fields()))).orElse(originalRowSize);
@@ -479,7 +482,8 @@ public class BigQueryDataSourceReaderContext {
 
             @Override
             public OptionalLong numRows() {
-              return OptionalLong.empty();
+              // Number of rows are only available for BigLake Managed tables.
+              return isBigLakeManagedTable ? OptionalLong.of(tableRowCount) : OptionalLong.empty();
             }
           };
       return tableStatisticsContext;